requirements.txt
python-socketio==5.7.2
opencv-python==4.8.0.74
numpy==1.26.0
Pillow==10.0.0
ultralytics==8.0.20
coordinate-calculator== 0.1.0
